backend/requirements.txt
asgiref==3.7.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==4.2.5
django-colorfield==0.7.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.2.3
django-filter==21.1
django-psycopg2-extension==0.0.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.2.1
drf-yasg==1.20.0
flake8==4.0.1
flake8-isort==4.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
psycopg2==2.8.6
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.6
reportlab==3.6.5
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.3
testfixtures==6.18.3
toml==0.10.2
typed-ast==1.4.3
typing_extensions==4.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
